Gateway integration scaffolding; general OCD
This commit introduces some Gateway integration scaffolding along with
some basic scraping capability. It doesn't complete the job, though.
There's also a fair bit of OCD cleanup going on.

At this point, the codeline is clean and ready for branching.

This is a "clean" checkpoint in development where all the code and stuff
is stable, tests pass, and there's not really much in here that isn't
getting used in some way. The exceptions are that Mojo is installed from
the cpanfile and there are few extra helper methods.

The plan now is to merge to master and nuke this branch, just for OCD
cleanup, so then to focus on QMA basic functionality duplication.