Fix "perfect spread" bug in build meet draw
If you have exactly the same number of teams to fill a set of rooms
(i.e. 18 teams in 6 rooms or 15 teams in 5 rooms), there was a weird
situation where a given quiz got repeated every so often. This commit
fixed that... but in a less than totally awesome way. The situation only
seems to happen when you have a "perfect spread" with 3 rooms, 5 rooms,
and 6 rooms. That's it. Everything else works fine.

It turns out the "solution" to this is to not prioritize team selection
based on a "has already been in this room" sort component. So this patch
turns off that specific sort component if and only if you happen to have
3, 5, or 6 rooms and are on a given set (or single time slice of
quizzes) where the error will show up.

I feel dirty.