Question type checking in overtime
Closes #248. Prior to this commit, numeric questions in overtime could
be of the same general/base type. (i.e. a "CVR" and "CR" could appear in
a set of 3 questions like 21, 22, and 23) With this commit, that in
theory can't happen unless near the end of the question queue, and if it
does happen, there's a warning with a recommendation to manually replace
the current/new question.